Add polynomial tests module and refactor tests into separate file
