+++ rings/src/parsing/mod.rs
@@ -1,1 +1,3 @@
 pub mod polynomial;
+#[cfg(test)]
+mod polynomial_tests;
